Added signal support (#29)
* Added SignalTypes, refactoring

* Attempt to support signal in given error message

* refactoring

* final refactoring

* Fixed tests

* Refactoring tests

* Added checking non optional fields

* Minor

* Fixed review issue

* Fixed tests

* Added required filed (sid) in requests of gatling scripts

* Changed streamId type to Long (Nonce.js)

* Changed scalecube-services version and jackson

* Minor

+++ Core/src/test/java/io/scalecube/gateway/core/TestInputs.java
@@ -1,7 +1,7 @@
 package io.scalecube.gateway.core;
 
 public interface TestInputs {
-  Integer SID = 42;
+  Long SID = 42L;
   Integer I = 423;
   Integer SIG = 422;
   String Q = "/test/test";
